src/requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
Django==3.2.9
django-bootstrap-v5==1.0.6
exif==1.3.2
importlib-metadata==3.10.1
Pillow==8.3.2
plum-py==0.7.0
psycopg2-binary==2.9.1
pytz==2021.3
soupsieve==2.2.1
sqlparse==0.4.2
zipp==3.6.0
